refactor(photos): extract PhotoGrid from PhotosTab
